(PC-24106)[PRO] feat: focus to error in change password forms
