+++ pro/src/pages/LostPassword/ChangePasswordRequestForm/ChangePasswordRequestForm.tsx
@@ -2,6 +2,7 @@ import { Form, useField } from 'formik'
 import React from 'react'
 
 import FormLayout from 'components/FormLayout'
+import { useScrollToFirstErrorAfterSubmit } from 'hooks'
 import { SubmitButton } from 'ui-kit'
 import { ButtonVariant } from 'ui-kit/Button/types'
 import EmailSpellCheckInput from 'ui-kit/form/EmailSpellCheckInput'
@@ -9,6 +10,7 @@ import EmailSpellCheckInput from 'ui-kit/form/EmailSpellCheckInput'
 import styles from './ChangePasswordRequestForm.module.scss'
 
 const ChangePasswordRequestForm = (): JSX.Element => {
+  useScrollToFirstErrorAfterSubmit()
   const [field] = useField('email')
 
   return (
