+++ pro/src/pages/ResetPassword/ChangePasswordForm/ChangePasswordForm.module.scss
@@ -2,10 +2,6 @@
 @use "styles/mixins/_fonts.scss" as fonts;
 
 .change-password-form {
-  h1 {
-    margin-bottom: rem.torem(16px);
-  }
-
   label {
     margin-bottom: rem.torem(16px);
   }
