--- pro/src/pages/ResetPassword/ChangePasswordForm/index.ts
@@ -1,1 +1,0 @@
-export { default } from './ChangePasswordForm'
