+++ pro/src/pages/ResetPassword/ResetPassword.module.scss
@@ -23,3 +23,7 @@
 .content {
   @include layout.full-content-side;
 }
+
+.change-password-title {
+  margin-bottom: rem.torem(16px);
+}
